Move folder.
