+++ src/tensor_ops/mod.rs
@@ -147,10 +147,10 @@ mod add;
 mod bce;
 mod boolean;
 mod broadcast_to;
-mod cat;
 mod choose;
 mod clamp;
 mod cmp;
+mod concat;
 mod cos;
 mod div;
 mod dropout;
